Move custom fields into separate file
