+++ front/app/containers/Admin/projects/project/analysis/InputPreview/components/CustomFields/LongFieldValue/CustomFieldLongFieldValues/PolygonLongField.tsx
